refactor: move warningballon out of the own folder
