+++ src/components/contact-with-me/comment-section/CommentSection.tsx
@@ -10,7 +10,7 @@ import {
 import { FlexColumnCenteredContainer } from "../../../components/styled-component/Container";
 // components
 import CommentComponent from "../comment-component/CommentComponent";
-import WarningBalloon from "../../../components/warning-balloon/WarningBalloon";
+import WarningBalloon from "../../WarningBalloon";
 
 export default function CommentSection() {
   const [inputUserName, setInputUserName] = useState<string>("");
